+++ nexusvault.format.m3/src/main/java/nexusvault/format/m3/v100/debug/Task.java
@@ -1,0 +1,5 @@
+package nexusvault.format.m3.v100.debug;
+
+public interface Task {
+	void runTask(Util debuger);
+}