+++ nexusvault.format.m3/docs/CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com)
 
 ## [Unreleased]
+### Added
+- Support to 'debug' a m3 model, this means to write as many values/structs to a table format, which can, for example, be dumped as csv
 
 ## [0.6.0] - 2019-01-28
 ### Added